Store money as Long cents, not Double
Currency was stored as Double across entities, prefs, and BalanceCalculator,
with the core halving mechanic minting fractional cents papered over by
roundToLong and a float-equality check. Money must never live in binary
floating point.

- Store money as Long minor units (cents) across CheckIn, CashOut,
  UserPreferences, DAOs, repositories, viewmodels, UI, and the widget.
- Replace balance halving's roundToLong dance with integer round-half-up
  ((balance + 1) / 2) and drop BalanceCalculator's rounding hack.
- Replace the float-equality recalculation guard with integer comparison.
- Add Formatters.parseDollarsToCents / centsToInput for the input boundary;
  formatCurrency now takes cents.
- Bump the Room schema to v2 (conversion folds into the multi-habit rewrite).

Closes #5

+++ app/src/main/java/com/oink/app/data/CashOutRepository.kt
@@ -36,7 +36,7 @@ class CashOutRepository(
     /**
      * Flow of total amount cashed out all-time.
      */
-    val totalCashedOut: Flow<Double> = cashOutDao.getTotalCashedOutFlow()
+    val totalCashedOut: Flow<Long> = cashOutDao.getTotalCashedOutFlow()
 
     /**
      * Cash out from the piggy bank! 🎉
@@ -56,7 +56,7 @@ class CashOutRepository(
      */
     suspend fun cashOut(
         name: String,
-        amount: Double,
+        amount: Long,
         emoji: String = "🎁"
     ): CashOut? {
         // Get ACTUAL current balance (after existing deductions)
@@ -97,7 +97,7 @@ class CashOutRepository(
      * Get the ACTUAL current balance after all deductions.
      * Uses BalanceCalculator for the actual calculation.
      */
-    private suspend fun getCurrentBalance(): Double {
+    private suspend fun getCurrentBalance(): Long {
         return BalanceCalculator.calculateActualBalance(
             checkInBalance = checkInRepository.getCurrentBalanceOnce(),
             totalCashedOut = cashOutDao.getTotalCashedOut(),
@@ -108,7 +108,7 @@ class CashOutRepository(
     /**
      * Get total amount cashed out all-time.
      */
-    suspend fun getTotalCashedOut(): Double {
+    suspend fun getTotalCashedOut(): Long {
         return cashOutDao.getTotalCashedOut()
     }
